+++ assets/js/index.js
@@ -10,13 +10,6 @@ require("jquery");
 require("bootstrap");
 require("bootstrap-icons/font/bootstrap-icons.css");
 require("bootstrap4-toggle");
-const $ = require("jquery");
-const {
-  setMessageCount,
-  setTaskProgress,
-  getMessageCount,
-  whenDone,
-} = require("./messages");
 const DarkMode = require("./darkMode");
 const registerServiceWorker = require("./serviceWorker");
 
@@ -33,9 +26,4 @@ document.addEventListener("DOMContentLoaded", () => {
 
 module.exports = {
   darkMode,
-  setMessageCount,
-  setTaskProgress,
-  getMessageCount,
-  $,
-  whenDone,
 };