+++ webpack.config.js
@@ -22,6 +22,7 @@ module.exports = {
   entry: {
     index: "./assets/js/index.js",
     highlight: "./assets/js/highlight.js",
+    messages: "./assets/js/messages.js",
     main_css: [
       "@fortawesome/fontawesome-free/css/all.css",
       "bootstrap/dist/css/bootstrap.css",