+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/jss/compare/v1.27.0...HEAD)
 ------------------------------------------------------------------------
 ### Changed
+- add messages bundle
 - add highlight bundle
 
 [1.27.0](https://github.com/jshwi/jss/releases/tag/v1.27.0) - 2023-07-26